requirements.txt
jmespath
alibaba-cloud-python-sdk-v2
happy-python
tencentcloud-sdk-python